+++ CommandLineToolSwift/CommandLineToolSwift/Account.swift
@@ -1,0 +1,15 @@
+//
+//  Account.swift
+//  CommandLineToolSwift
+//
+//  Created by Mateus Campos on 11/04/16.
+//  Copyright © 2016 Mateus Campos. All rights reserved.
+//
+
+import Foundation
+
+
+class Account {
+    var account:Int? = nil
+    var ammount:Int? = nil
+}
