+++ CommandLineToolSwift/CommandLineToolSwift/Path.swift
@@ -1,0 +1,14 @@
+//
+//  Path.swift
+//  CommandLineToolSwift
+//
+//  Created by Mateus Campos on 11/04/16.
+//  Copyright © 2016 Mateus Campos. All rights reserved.
+//
+
+import Foundation
+
+
+class Path {
+    var filePath:String? = nil
+}
