+++ CommandLineToolSwift/CommandLineToolSwift/Transaction.swift
@@ -1,0 +1,15 @@
+//
+//  Transaction.swift
+//  CommandLineToolSwift
+//
+//  Created by Mateus Campos on 11/04/16.
+//  Copyright © 2016 Mateus Campos. All rights reserved.
+//
+
+import Foundation
+
+
+class Transaction {
+    var account:Int? = nil
+    var transaction:Int? = nil
+}
