+++ CommandLineToolSwift/resources/accs.csv
@@ -1,0 +1,1 @@
+123,12345,456,11111,789,54321
