+++ CommandLineToolSwift/resources/transactions.csv
@@ -1,0 +1,1 @@
+123,22222,456,-54321,789,44444
